исправил бэк

+++ backend/src/infrastructure/database/models.py
@@ -1,6 +1,6 @@
 from datetime import datetime
 
-from sqlalchemy import Column, Date, DateTime, ForeignKey, Index, Integer, String, Text
+from sqlalchemy import BigInteger, Column, Date, DateTime, ForeignKey, Index, Integer, String, Text
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 
@@ -71,5 +71,5 @@ class PanelAccessModel(Base):
     __tablename__ = "panel_access"
 
     id = Column(Integer, primary_key=True, index=True)
-    user_id = Column(Integer, nullable=False, index=True)
+    user_id = Column(BigInteger, nullable=False, index=True)
     accessed_at = Column(DateTime, default=datetime.utcnow, nullable=False)
